# docs/api/external-v1.openapi.yaml
openapi: 3.1.0
info:
  title: bluDM External Authoring API
  version: 1.0.0
  description: Versioned REST adapter over the same application services used by bluDM MCP.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://{host}
    description: Replace the host with the public or private bluDM deployment hostname.
    variables:
      host:
        default: bludm.invalid
security:
  - bearerAuth: []
...
paths:
  /api/external/v1/campaigns:
    get:
      operationId: listCampaigns
      summary: List campaigns available to the credential.
      tags: [discovery]
      parameters: [{ $ref: "#/components/parameters/Limit" }, { $ref: "#/components/parameters/Cursor" }]
      responses: { "200": { $ref: "#/components/responses/CampaignPage" }, default: { $ref: "#/components/responses/Error" } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Scoped bludm_v1_ API token or configured OIDC JWT access token.
  parameters:
...
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
    Cursor: { name: cursor, in: query, schema: { type: string }, description: Opaque cursor from the previous page. }
...
  responses:
    CampaignPage: { description: Campaign discovery page., content: { application/json: { schema: { $ref: "./external-v1.schemas.yaml#/components/schemas/CampaignPage" } } } }
...
    Error:
      description: Typed external API error.
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            error: { code: conflict, message: encounter revision changed, requestId: req-123 }
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { enum: [unauthorized, forbidden, not_found, validation_failed, conflict, idempotency_conflict, rate_limited, timeout, unsupported, internal_error] }
            message: { type: string }
            details: { type: object }
            requestId: { type: string }
          additionalProperties: false
      additionalProperties: false
...